Improvements

+++ backend/create.js
@@ -6,7 +6,8 @@ let server = fs.readFileSync(path.resolve(__dirname, './files/server.js'), 'utf8
 let enzoCreateEndpoints = fs.readFileSync(path.resolve(__dirname, './files/enzoCreateAPI.js'), 'utf8')
 
 let routes = `const express = require('express')\nconst r = express.Router()\nmodule.exports = r`
-
+let enzoEndpointTemplate = fs.readFileSync(path.resolve(__dirname, './templates/enzoEndpointTemplate.js'), 'utf8')
+let enzoModelTemplate = fs.readFileSync(path.resolve(__dirname, './templates/enzoModelTemplate.js'), 'utf8')
 
 let backendOnly = () => {
   if (fs.existsSync('./package.json')) {
@@ -20,13 +21,30 @@ let backendOnly = () => {
     })
     // need to see if enzo exists, if not make folder sync then create this file
     if (fs.existsSync('./enzo')) {
-      fs.writeFile(`./enzo/api.js`, enzoCreateEndpoints, (err) => {
-        if (err) throw err
-      })
+      if (fs.existsSync('./enzo/api.js')) {
+        
+      } else {
+        fs.writeFile(`./enzo/api.js`, enzoCreateEndpoints, (err) => {
+          if (err) console.error(err)
+        })
+        fs.writeFile(`./enzo/templates/enzoEndpointTemplate.js`, enzoEndpointTemplate, (err) => {
+          if (err) console.error(err)
+        })
+        fs.writeFile(`./enzo/templates/enzoModelTemplate.js`, enzoModelTemplate, (err) => {
+          if (err) console.error(err)
+        })
+      }
     } else {
       fs.mkdirSync('./enzo')
       fs.writeFile(`./enzo/api.js`, enzoCreateEndpoints, (err) => {
         if (err) throw err
+      })
+      fs.mkdirSync('./enzo/templates')
+      fs.writeFile(`./enzo/templates/enzoEndpointTemplate.js`, enzoEndpointTemplate, (err) => {
+        if (err) console.error(err)
+      })
+      fs.writeFile(`./enzo/templates/enzoModelTemplate.js`, enzoModelTemplate, (err) => {
+        if (err) console.error(err)
       })
     }
